+++ console_unix.go
@@ -278,7 +278,7 @@ func GetCursorPos() (int,int) {
 }
 
 /// get an input string from stdin, in raw mode
-func getInput(evalfs uint64, prompt string, pane string, row int, col int, pcol string, histEnable bool, hintEnable bool, mask string) (s string, eof bool, broken bool) {
+func getInput(evalfs uint32, prompt string, pane string, row int, col int, pcol string, histEnable bool, hintEnable bool, mask string) (s string, eof bool, broken bool) {
 
     sprompt := sparkle(prompt)
